% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ijlyttle_github_io.R
\name{ijlyttle_github_io}
\alias{ijlyttle_github_io}
\title{html_document() for ijlyttle_github_io}
\usage{
ijlyttle_github_io(...)
}
\arguments{
\item{...}{arguments passed to
\code{rmarkdown::\link[rmarkdown]{html_document}}}
}
\value{
R Markdown output format to pass to
  \code{rmarkdown::\link[rmarkdown]{render}}
}
\description{
html_document() for ijlyttle_github_io
}
\examples{
\dontrun{
  rmarkdown::render("foo.Rmd", ijlyttle_github_io())
}
}
\seealso{
\url{http://rmarkdown.rstudio.com/developer_custom_formats.html}
}

